fix: Fix the logic of epoll event handling
I misunderstood the EPOLLOUT event flag meaning.
For the socket fd which created by accept(), the EPOLLOUT means the socket is writable, the OS kernel buffer is not full and can accept more data to write.
And set the EPOLLOUT flag for the listening socket fd won't trigger the write event.