chore: Use all CPU cores for compilation
This speeds up compilation significantly.

The `MAKEFLAGS` is a default variable of Make. Whatever Make option you assign to it is the same as if you would run `make` with that option.
Assigning the `-j` option lets Make always run with that option, even when it wasn't given in the command line.
`-j` stands for "jobs" (`--jobs`).

The `re` target had to be slightly changed so that its calls to `fclean` and `all` don't run in parallel. Having `fclean` as a prerequisite and then calling `all` in the recipe ensures that.